refactor: extract fleet dispatch and credentials into bounded FastAPI routers
Addresses issue #4 — reduces server.py coupling by extracting two well-bounded
API domains into their own APIRouter modules under backend/routers/:

- routers/dispatch.py: /api/fleet/dispatch/* endpoints (list, validate, submit)
  — depends only on dispatch_contract; no global server state
- routers/credentials.py: /api/credentials endpoint
  — reads os.environ and probes local binaries; no GitHub client dependency

server.py registers these via app.include_router() and removes ~270 lines of
inline endpoint code. Remaining endpoint domains are candidates for future
extraction (tracked in issue #4).

Adds test_backend_routers.py to assert the routers are correctly structured,
expose the expected paths, and that server.py uses include_router (not inline).

Updates SPEC.md section 2.1 to document the bounded routers pattern and the
ongoing migration plan.

Closes #4
